requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Machine Learning
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Data Processing
yfinance==0.2.28
alpaca-trade-api==3.1.1
requests==2.31.0

# Time Series
statsmodels==0.14.0
ta==0.11.0

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Utilities
python-dotenv==1.0.0
python-json-logger==2.0.7
schedule==1.2.0
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

